requirements.txt
agno
python-dotenv
anthropic
pydantic
requests
pandas
openai
voyageai
fastapi
uvicorn
python-multipart
python-docx